Refactor inspector fields & renderer background
Replace Speed with Value for FloatField/IntField and update all inspector usages; make Renderer.Clear accept an optional Color and use the current scene camera color in Editor and GameLoop; add Camera.Color and constructors to support background color. Fix ImGui TextInput Marshal usage and harden several inspector UI flows (list items, pickers, popups). Miscellaneous whitespace/usings and small cleanup across multiple files.

+++ KoreEngine.Editor/ViewportPanel.cs
@@ -129,7 +129,7 @@ namespace KoreEngine.Editor
             var bounds = collider.Bounds;
 
             byte r = collider.IsTrigger ? (byte)255 : (byte)0;
-            byte g = (byte)255;
+            byte g = 255;
             byte b = collider.IsTrigger ? (byte)0 : (byte)0;
 
             renderer.DrawRectOutline(
